tools.py
## Importing libraries and files
import os
from typing import Optional, Type
from crewai.tools import BaseTool
from pydantic import BaseModel, Field

class SearchInput(BaseModel):
    """Input for the search tool."""
    query: str = Field(..., description="The search query")

class SimpleSearchTool(BaseTool):
    name: str = "simple_search_tool"
    description: str = "A simple search tool that can be used to search the web"
    args_schema: Type[BaseModel] = SearchInput
    
    def _run(self, query: str) -> str:
        # This is a placeholder implementation
        return f"Search results for: {query}\nThis is a placeholder search result. In a real implementation, this would search the web."

class PDFInput(BaseModel):
    """Input for the PDF tool."""
    file_path: str = Field(..., description="The path to the PDF file")

class SimplePDFTool(BaseTool):
    name: str = "simple_pdf_tool"
    description: str = "A simple tool to read text from PDF files"
    args_schema: Type[BaseModel] = PDFInput
    
    def _run(self, file_path: str) -> str:
        try:
            if not os.path.exists(file_path):
                return f"Error: File not found at {file_path}"
            
            # Simple text file reading as a fallback
            if file_path.endswith('.txt'):
                with open(file_path, 'r', encoding='utf-8') as f:
                    return f.read()
            else:
                return f"This is a placeholder for PDF content from {file_path}. In a real implementation, this would extract text from the PDF."
        except Exception as e:
            return f"Error reading file: {str(e)}"

# Create instances of the tools
search_tool = SimpleSearchTool()
pdf_tool = SimplePDFTool()