requirements-minimal.txt
crewai
fastapi
python-multipart
python-dotenv
langchain-google-genai
google-generativeai
crewai-tools
